refactor and bugfixes
