+++ surge/nSurge.go
@@ -6,7 +6,6 @@ import (
 	"math/rand"
 	"net"
 	"os"
-	"os/user"
 	"runtime"
 	"strconv"
 	"strings"
@@ -18,29 +17,12 @@ import (
 
 	bitmap "github.com/boljen/go-bitmap"
 	movavg "github.com/mxmCherry/movavg"
-	nkn "github.com/nknorg/nkn-sdk-go"
 	"github.com/wailsapp/wails"
 )
 
 // SurgeActive is true when client is operational
 var SurgeActive bool = false
-
-//ChunkSize is size of chunk in bytes (256 kB)
-const ChunkSize = 1024 * 256
-
-//NumClients is the number of NKN clients
-const NumClients = 8
-
-//NumWorkers is the total number of concurrent chunk fetches allowed
-const NumWorkers = 32
-
-const remotePath = "downloads"
-
-var remoteFolder = ""
-
 var subscribers []string
-
-var clientInitialized = false
 
 //OS folder permission bitflags
 const (
@@ -81,8 +63,6 @@ var sendSize int64
 var receivedSize int64
 
 var startTime = time.Now()
-
-var client *nkn.MultiClient
 
 var clientOnlineMap map[string]int64
 
@@ -220,114 +200,35 @@ func SetVisualMode(visualMode int) {
 
 // Start initializes surge
 func Start(args []string) {
-	var err error
 
-	var dirFileMode os.FileMode
-	var dir = GetSurgeDir()
-	dirFileMode = os.ModeDir | (osUserRwx | osAllR)
+	//Initialize all our global data maps
+	clientOnlineMap = make(map[string]int64)
+	downloadBandwidthAccumulator = make(map[string]int)
+	uploadBandwidthAccumulator = make(map[string]int)
+	zeroBandwidthMap = make(map[string]bool)
+	fileBandwidthMap = make(map[string]BandwidthMA)
+	chunksInTransit = make(map[string]bool)
 
-	myself, err := user.Current()
-	if err != nil {
-		pushError("Error on startup", err.Error())
-	}
-	homedir := myself.HomeDir
-
-	remoteFolder = homedir + string(os.PathSeparator) + "Downloads" + string(os.PathSeparator) + "surge_" + remotePath
-
-	if _, err := os.Stat(dir); os.IsNotExist(err) {
+	//Initialize folder structures on os filesystem
+	newlyCreated := InitializeFolders()
+	if newlyCreated {
 		// seems like this is the first time starting the app
 		//set tour to active
 		DbWriteSetting("Tour", "true")
 		//set default mode to light
 		DbWriteSetting("DarkMode", "false")
-
-		os.Mkdir(dir, dirFileMode)
 	}
 
-	//Ensure remote folders exist
-	if _, err := os.Stat(remoteFolder); os.IsNotExist(err) {
-		os.Mkdir(remoteFolder, dirFileMode)
+	//Initialize our surge nkn client
+	InitializeClient()
+
+	//Insert new file from arguments and start download
+	if args != nil && len(args) > 0 && len(args[0]) > 0 {
+		platform.AskUser("startDownloadMagnetLinks", "{files : ["+args[0]+"]}")
 	}
 
-	account := InitializeAccount()
-	client, err = nkn.NewMultiClient(account, "", NumClients, false, nil)
-
-	clientInitialized = true
-
-	if err != nil {
-		pushError("Error on startup", err.Error())
-	} else {
-		log.Println("MY ADDRESS:", client.Addr().String())
-		<-client.OnConnect.C
-
-		pushNotification("Client Connected", "Successfully connected to the NKN network")
-
-		client.Listen(nil)
-		SurgeActive = true
-		go Listen()
-
-		topicEncoded := TopicEncode(TestTopic)
-
-		clientOnlineMap = make(map[string]int64)
-		downloadBandwidthAccumulator = make(map[string]int)
-		uploadBandwidthAccumulator = make(map[string]int)
-		zeroBandwidthMap = make(map[string]bool)
-		fileBandwidthMap = make(map[string]BandwidthMA)
-		chunksInTransit = make(map[string]bool)
-
-		dbFiles := dbGetAllFiles()
-		var filesOnDisk []File
-
-		for i := 0; i < len(dbFiles); i++ {
-			if FileExists(dbFiles[i].Path) {
-				filesOnDisk = append(filesOnDisk, dbFiles[i])
-			} else {
-				dbFiles[i].IsMissing = true
-				dbFiles[i].IsDownloading = false
-				dbFiles[i].IsUploading = false
-				dbInsertFile(dbFiles[i])
-			}
-		}
-
-		go BuildSeedString(filesOnDisk)
-		for i := 0; i < len(filesOnDisk); i++ {
-			go restartDownload(filesOnDisk[i].FileHash)
-		}
-
-		go sendSeedSubscription(topicEncoded, "Surge File Seeder")
-		go GetSubscriptions(topicEncoded)
-
-		go updateGUI()
-
-		go queryRemoteForFiles()
-
-		go platform.WatchOSXHandler()
-
-		go rescanPeers()
-
-		//Insert new file from arguments and start download
-		if args != nil && len(args) > 0 && len(args[0]) > 0 {
-			platform.AskUser("startDownloadMagnetLinks", "{files : ["+args[0]+"]}")
-		}
-
-		//Just paste one of your own magnets (from the startup logs) here to download something over nkn from yourself to test if no-one is online
-		//go ParsePayloadString("surge://|file|justatvshow.mp4|219091405|cd0731496277102a869dacb0e99b7708c2b708824b647ffeb267de4743b7856e|a536528d2e321623375535af88974d7a7899836f9b84644320023bc3af3b9cf1|/")
-	}
-}
-
-//Stop cleanup for surge
-func Stop() {
-	client.Close()
-	client = nil
-}
-
-func rescanPeers() {
-	defer RecoverAndLog()
-	for true {
-		time.Sleep(time.Minute)
-		topicEncoded := TopicEncode(TestTopic)
-		go GetSubscriptions(topicEncoded)
-	}
+	//Run gui update worker for frontend
+	go updateGUI()
 }
 
 func updateClientOnlineMap() {
@@ -382,9 +283,18 @@ func queryRemoteForFiles() {
 	}
 }
 
-//GetNumberOfRemoteClient returns number of clients and online clients
-func GetNumberOfRemoteClient() (int, int) {
-	return numClientsSubscribed, numClientsOnline
+func chunkMapFull(s []byte, num int) bool {
+	//No chunkmap means no download was initiated, all chunks are local
+	if s == nil {
+		return true
+	}
+
+	for i := 0; i < num; i++ {
+		if bitmap.Get(s, i) == false {
+			return false
+		}
+	}
+	return true
 }
 
 func updateGUI() {
@@ -404,25 +314,29 @@ func updateGUI() {
 
 			fileProgressMap[session.FileHash] = float32(float64(session.Downloaded) / float64(session.FileSize))
 
-			if session.Downloaded == session.FileSize {
-
-				filename := "unknown"
-				listedFileByHash := getListedFileByHash(session.FileHash)
-				if listedFileByHash != nil {
-					filename = listedFileByHash.FileName
-				}
-				platform.ShowNotification("Download Finished", "Download for "+filename+" finished!")
-				pushNotification("Download Finished", filename)
-				session.session.Close()
-
+			if session.Downloaded >= session.FileSize {
 				fileEntry, err := dbGetFile(session.FileHash)
 				if err != nil {
 					pushError("Error on download complete", err.Error())
+					continue
 				}
-				fileEntry.IsDownloading = false
-				fileEntry.IsUploading = true
-				dbInsertFile(*fileEntry)
-				go AddToSeedString(*fileEntry)
+
+				fmt.Println(fileEntry.ChunkMap)
+				fmt.Println(fileEntry.ChunkMap)
+				fmt.Println(fileEntry.ChunkMap)
+				fmt.Println(fileEntry.ChunkMap)
+				fmt.Println(fileEntry.ChunkMap)
+
+				if chunkMapFull(fileEntry.ChunkMap, fileEntry.NumChunks) {
+					platform.ShowNotification("Download Finished", "Download for "+fileEntry.FileName+" finished!")
+					pushNotification("Download Finished", fileEntry.FileName)
+					session.session.Close()
+
+					fileEntry.IsDownloading = false
+					fileEntry.IsUploading = true
+					dbInsertFile(*fileEntry)
+					go AddToSeedString(*fileEntry)
+				}
 			}
 		}
 		sessionsWriteLock.Unlock()
@@ -495,33 +409,6 @@ func fileBandwidth(FileID string) (Download int, Upload int) {
 	fileBandwidthMap[FileID].Upload.Add(float64(upAccu))
 
 	return int(fileBandwidthMap[FileID].Download.Avg()), int(fileBandwidthMap[FileID].Upload.Avg())
-
-	//Take bandwith delta
-	/*deltaDownload := int(Session.Downloaded - Session.deltaDownloaded)
-	Session.deltaDownloaded = Session.Downloaded
-	deltaUpload := int(Session.Uploaded - Session.deltaUploaded)
-	Session.deltaUploaded = Session.Uploaded
-
-	//Append to queue
-	Session.bandwidthDownloadQueue = append(Session.bandwidthDownloadQueue, deltaDownload)
-	Session.bandwidthUploadQueue = append(Session.bandwidthUploadQueue, deltaUpload)
-
-	//Dequeue if queue > 10
-	if len(Session.bandwidthDownloadQueue) > 10 {
-		Session.bandwidthDownloadQueue = Session.bandwidthDownloadQueue[1:]
-		Session.bandwidthUploadQueue = Session.bandwidthUploadQueue[1:]
-	}
-
-	var downloadMA10 = 0
-	var uploadMA10 = 0
-	for i := 0; i < len(Session.bandwidthDownloadQueue); i++ {
-		downloadMA10 += Session.bandwidthDownloadQueue[i]
-		uploadMA10 += Session.bandwidthUploadQueue[i]
-	}
-	downloadMA10 /= len(Session.bandwidthDownloadQueue)
-	uploadMA10 /= len(Session.bandwidthUploadQueue)
-
-	return downloadMA10, uploadMA10*/
 }
 
 //ByteCountSI converts filesize in bytes to human readable text
@@ -546,39 +433,6 @@ func getFileSize(path string) (size int64) {
 	}
 	// get the size
 	return fi.Size()
-}
-
-func sendSeedSubscription(Topic string, Payload string) {
-	defer RecoverAndLog()
-	txnHash, err := client.Subscribe("", Topic, 4320, Payload, nil)
-	if err != nil {
-		log.Println("Probably already subscribed", err)
-	} else {
-		log.Println("Subscribed: ", txnHash)
-	}
-}
-
-//GetSubscriptions .
-func GetSubscriptions(Topic string) {
-	defer RecoverAndLog()
-	subResponse, err := client.GetSubscribers(Topic, 0, 100, true, true)
-	if err != nil {
-		pushError("Error on get subscriptions", err.Error())
-		return
-	}
-
-	for k, v := range subResponse.SubscribersInTxPool.Map {
-		subResponse.Subscribers.Map[k] = v
-	}
-
-	subscribers = []string{}
-	for k, v := range subResponse.Subscribers.Map {
-		if len(v) > 0 {
-			if k != client.Addr().String() {
-				subscribers = append(subscribers, k)
-			}
-		}
-	}
 }
 
 // Stats .
@@ -852,27 +706,6 @@ func DownloadFile(Hash string) bool {
 	return true
 }
 
-func pushNotification(title string, text string) {
-	//If wails frontend is not yet binded, we wait in a task to not block main thread
-	if wailsRuntime == nil {
-
-		waitAndPush := func() {
-			for wailsRuntime == nil {
-				time.Sleep(50)
-			}
-			wailsRuntime.Events.Emit("notificationEvent", title, text)
-		}
-		go waitAndPush()
-	} else {
-		wailsRuntime.Events.Emit("notificationEvent", title, text)
-	}
-}
-
-func pushError(title string, text string) {
-	//log.Println("Emitting Event: ", "errorEvent", title, text)
-	wailsRuntime.Events.Emit("errorEvent", title, text)
-}
-
 //SearchQueryResult is a paging query result for file searches
 type SearchQueryResult struct {
 	Result []FileListing
@@ -904,17 +737,6 @@ func SearchFile(Query string, Skip int, Take int) SearchQueryResult {
 			}
 
 			tracked, err := dbGetFile(result.FileHash)
-			//if err == nil && tracked != nil {
-			//	result.IsTracked = true
-			//	result.IsAvailable = true
-
-			//If any chunk is missing set available to false
-			//	for i := 0; i < result.NumChunks; i++ {
-			//		if bitmap.Get(tracked.ChunkMap, i) == false {
-			//			result.IsAvailable = false
-			//			break
-			//		}
-			//}
 
 			//only add non-local files to the result
 			if err != nil && tracked == nil {
